+++ framework/src/main/kotlin/me/wieku/framework/time/IClock.kt
@@ -1,7 +1,7 @@
 package me.wieku.framework.time
 
 interface IClock {
-    var currentTime: Float
-    var clockRate: Float
+    var currentTime: Double
+    var clockRate: Double
     var isRunning: Boolean
 }
